+++ src/Time/Format.hs
@@ -31,12 +31,17 @@ module Time.Format
   , localTimeParseE
   ) where
 
-import           Data.Char ( isDigit, ord )
+import           Data.Char ( isDigit, isSpace, ord )
 import           Data.Int ( Int64 )
+import           Time.Calendar ( getDayOfTheYear )
 import           Time.Internal ( dateTimeFromUnixEpochP )
 import           Time.LocalTime
                    ( LocalTime (..), localTime, localTimeToGlobal )
-import           Time.Time ( Timeable (..), timeGetDateTimeOfDay )
+import           Time.Time
+                   ( Period (..), Timeable (..), dateAddPeriod
+                   , timeGetDateTimeOfDay
+                   )
+import           Time.Timezone ( Timezone (..), TimezoneMinutes (..) )
 import           Time.Utils ( pad2, pad4, padN )
 import           Time.Types
                    ( Date (..), DateTime (..), Elapsed (..), ElapsedP (..)
@@ -76,7 +81,7 @@ data TimeFormatElem =
     -- (1970-01-01 00:00:00 UTC).
   | Format_MilliSecond
     -- ^ The millisecond component only, padded to 3 characters (@000@ to
-    -- @999@). See 'Format_MircoSecond' and 'Format_NanoSecond' for other
+    -- @999@). See 'Format_MicroSecond' and 'Format_NanoSecond' for other
     -- named sub-second components.
   | Format_MicroSecond
     -- ^ The microseconds component only, padded to 3 characters (@000@ to
@@ -90,22 +95,24 @@ data TimeFormatElem =
     -- ^ Sub-second display with a precision of n digits, with n between @1@
     -- and @9@.
   | Format_TimezoneName
-    -- ^ Timezone name (e.g. GMT, PST). Not yet implemented.
+    -- ^ Timezone name.
   | Format_TzHM_Colon_Z
     -- ^ Zero UTC offset (@Z@) or timezone offset with colon (for example,
-    -- @+02:00@).
+    -- @02:00@, @+02:00@ or @-02:00@).
   | Format_TzHM_Colon
-    -- ^ Timezone offset with colon (for example, @+02:00@).
+    -- ^ Timezone offset with colon (for example, @02:00@, @+02:00@ or
+    -- @-02:00@).
   | Format_TzHM
-    -- ^ Timezone offset without colon (for example, @+0200@).
+    -- ^ Timezone offset without colon (for example, @0200@, @+0200@ or
+    -- @-0200@).
   | Format_Tz_Offset
-    -- ^ Timezone offset in minutes.
+    -- ^ Timezone offset in minutes (for example, @120@, @+120@ or @-120@).
   | Format_Spaces
     -- ^ One or more space-like characters.
   | Format_Text Char
     -- ^ A verbatim character.
   | Format_Fct TimeFormatFct
-    -- ^ Not implemented.
+    -- ^ A custom time format function. See v'TimeFormatFct'.
   deriving (Eq, Show)
 
 -- | Type representing format functions.
@@ -168,36 +175,21 @@ newtype TimeFormatString = TimeFormatString [TimeFormatElem]
 -- [@ns@]:    'Format_NanoSecond'. The nanoseconds component only, padded to 3
 --            characters (@000@ to @999@). See @ms@ and @us@/@μ@ for other named
 --            sub-second components.
--- [@p1@]:    'Format_Precision' @1@. Sub-second display with a precision of 1
---            digit.
--- [@p2@]:    'Format_Precision' @2@. Sub-second display with a precision of 2
---            digits.
--- [@p3@]:    'Format_Precision' @3@. Sub-second display with a precision of 3
---            digits.
--- [@p4@]:    'Format_Precision' @4@. Sub-second display with a precision of 4
---            digits.
--- [@p5@]:    'Format_Precision' @5@. Sub-second display with a precision of 5
---            digits.
--- [@p6@]:    'Format_Precision' @6@. Sub-second display with a precision of 6
---            digits.
--- [@p7@]:    'Format_Precision' @7@. Sub-second display with a precision of 7
---            digits.
--- [@p8@]:    'Format_Precision' @8@. Sub-second display with a precision of 8
---            digits.
--- [@p9@]:    'Format_Precision' @9@. Sub-second display with a precision of 9
---            digits.
+-- [@p\<n\>@]: 'Format_Precision' @\<n\>@. Sub-second display with a precision
+--             of @\<n\>@ digit(s), where @\<n\>@ is @1@ to @9@.
 -- [@TZH:M@]: 'Format_TzHM_Colon'. Timezone offset with colon (for example,
---            @+02:00@).
+--            @02:00@, @+02:00@ or @-02:00@).
 -- [@TZHM@]:  'Format_TzHM'. Timezone offset without colon (for example,
---            @+0200@).
--- [@TZOFS@]: 'Format_Tz_Offset'. Timezone offset in minutes.
--- [@\<space\>@]: 'Format_Spaces'. One or more space-like characters.
--- [@\\\\\<character\>@]: 'Format_Text' @\<character\>@. A verbatim character.
--- [@\<character\>@]:     'Format_Text' @\<character\>@. A verbatim character.
+--            @0200@, @+0200@ or @-0200@).
+-- [@TZOFS@]: 'Format_Tz_Offset'. Timezone offset in minutes (for example,
+--            @120@, @+120@ or @-120@).
+-- [@\<space\>@]:       'Format_Spaces'. One or more space-like characters.
+-- [@\\\<character\>@]: 'Format_Text' @\<character\>@. A verbatim character.
+-- [@\<character\>@]:   'Format_Text' @\<character\>@. A verbatim character.
 --
 -- For example:
 --
--- >>> let mDateTime = timeParse (toFormat "ms \\ms us \\us ns \\ns") "123 ms 456 us 789 ns"
+-- >>> let mDateTime = timeParse ("ms \\ms us \\us ns \\ns") "123 ms 456 us 789 ns"
 -- >>> timeGetNanoSeconds <$> mDateTime
 -- Just 123456789ns
 --
@@ -265,16 +257,17 @@ instance TimeFormat String where
 
 instance TimeFormat ISO8601_Date where
   toFormat _ =
-    TimeFormatString [Format_Year,dash,Format_Month2,dash,Format_Day2]
+    TimeFormatString [Format_Year, dash, Format_Month2, dash, Format_Day2]
    where
     dash = Format_Text '-'
 
 instance TimeFormat ISO8601_DateAndTime where
   toFormat _ = TimeFormatString
-    [ Format_Year,dash,Format_Month2,dash,Format_Day2 -- date
+    [ Format_Year, dash, Format_Month2, dash, Format_Day2 -- date
     , Format_Text 'T'
-    , Format_Hour,colon,Format_Minute,colon,Format_Second -- time
-    , Format_TzHM_Colon_Z -- zero UTC offset (Z) or timezone offset with colon +HH:MM
+    , Format_Hour, colon, Format_Minute, colon, Format_Second -- time
+    , Format_TzHM_Colon_Z
+      -- Either timezone offset with colon (±HH:MM) or UTC zero offset (Z).
     ]
    where
     dash = Format_Text '-'
@@ -311,6 +304,7 @@ printWith fmt t tzOfs@(TimezoneOffset tz) = concatMap fmtToString fmtElems
   fmtToString Format_Month2   = pad2 (fromEnum (dateMonth date) + 1)
   fmtToString Format_Month    = show (fromEnum (dateMonth date) + 1)
   fmtToString Format_MonthName_Short = take 3 $ show (dateMonth date)
+  fmtToString Format_DayYear  = show (getDayOfTheYear date)
   fmtToString Format_Day2     = pad2 (dateDay date)
   fmtToString Format_Day      = show (dateDay date)
   fmtToString Format_Hour     = pad2 (fromIntegral (todHour tm) :: Int)
@@ -323,7 +317,7 @@ printWith fmt t tzOfs@(TimezoneOffset tz) = concatMap fmtToString fmtElems
       | n >= 1 && n <= 9 = padN n (ns `div` (10 ^ (9 - n)))
       | otherwise        = error "invalid precision format"
   fmtToString Format_UnixSecond = show unixSecs
-  fmtToString Format_TimezoneName   = "" --
+  fmtToString Format_TimezoneName = timezoneName $ TimezoneMinutes tz
   fmtToString Format_Tz_Offset = show tz
   fmtToString Format_TzHM = show tzOfs
   fmtToString Format_TzHM_Colon_Z
@@ -335,15 +329,32 @@ printWith fmt t tzOfs@(TimezoneOffset tz) = concatMap fmtToString fmtElems
        in sign ++ pad2 tzH ++ ":" ++ pad2 tzM
   fmtToString Format_Spaces   = " "
   fmtToString (Format_Text c) = [c]
-  fmtToString f = error ("unimplemented printing format: " ++ show f)
+  fmtToString (Format_Fct tff) = timeFormatPrint tff dateTime tzOfs
 
   (TimeFormatString fmtElems) = toFormat fmt
 
   (Elapsed (Seconds unixSecs)) = timeGetElapsed t
-  (DateTime date tm) = timeGetDateTimeOfDay t
+  dateTime@(DateTime date tm) = timeGetDateTimeOfDay t
   (NanoSeconds ns) = timeGetNanoSeconds t
 
 -- | Given the specified format, pretty print the given local time.
+--
+-- A v'Format_TimezoneName' will print using the ±HH:MM format, where @0@
+-- offset is printed as @-00:00@.
+--
+-- A v'Format_TzHM' will print using the ±HHMM format, where @0@ offset is
+-- printed as @+0000@.
+--
+-- A v'Format_TzHM_Colon' will print using the ±HH:MM format, where @0@ offset
+-- is printed as @+00:00@.
+--
+-- A v'Format_TzHM_Colon_Z' will print using the ±HH:MM format, but where @0@
+-- offset is printed as @Z@.
+--
+-- A v'Format_Tz_Offset' will print non-negative offsets without using an
+-- initial @+@.
+--
+-- A v'Format_Spaces' will print a single space character.
 localTimePrint ::
      (TimeFormat format, Timeable t)
   => format      -- ^ The format to use for printing.
@@ -352,7 +363,8 @@ localTimePrint ::
 localTimePrint fmt lt =
   printWith fmt (localTimeUnwrap lt) (localTimeGetTimezone lt)
 
--- | Given the specified format, pretty print the given time.
+-- | Like 'localTimePrint' but the time zone of the time to print will be taken
+-- to be UTC.
 timePrint ::
      (TimeFormat format, Timeable t)
   => format -- ^ The format to use for printing.
@@ -360,11 +372,49 @@ timePrint ::
   -> String
 timePrint fmt t = printWith fmt t timezone_UTC
 
--- | Given the specified format, try to parse the given string as time value.
+-- | Given the specified format, try to parse the given string as
+-- a t'LocalTime' t'DateTime' value.
 --
--- On failure, the parsing function returns the reason of the failure.
+-- On failure, yields a 'Left' value with a pair of the
+-- current t'TimeFormatElem' value and the reason for the failure.
 --
--- If successful, yield the parsed value and the remaining unparsed string.
+-- If successful, yields a 'Right' value with a pair of the parsed value and
+-- the remaining unparsed string.
+--
+-- The default parsed t'LocalTime' t'DateTime' value is \'all zeros'\. For
+-- example:
+--
+-- >>> let zeroDate = Date 0 January 0
+-- >>> let zeroTime = TimeOfDay 0 0 0 0
+-- >>> let zeroLocalTime = localTime timezone_UTC (DateTime zeroDate zeroTime)
+-- >>> localTimeParseE "" "" == Right (zeroLocalTime, "")
+-- True
+--
+-- Later t'TimeFormatElem' values can modify the result of earlier
+-- t'TimeFormatElem' values. For example:
+--
+-- >>> let toYear = dateYear . dtDate . localTimeUnwrap . fst
+-- >>> toYear <$> (localTimeParseE "YYYY YYYY" "2025 2024")
+-- Right 2024
+--
+-- A v'Format_DayYear' interprets the day of year based on the previously parsed
+-- year or, by default, a leap year. For example:
+--
+-- >>> let toMonth = dateMonth . dtDate . localTimeUnwrap . fst
+-- >>> let format1 = [Format_Year4, Format_Spaces, Format_DayYear]
+-- >>> let format2 = [Format_DayYear, Format_Spaces, Format_Year4]
+-- >>> toMonth <$> (localTimeParseE format1 "2025 60")
+-- Right March
+-- >>> toMonth <$> (localTimeParseE format2 "60 2025")
+-- Right February
+--
+-- A v'Format_TimezoneName' will parse one or more non-white space characters
+-- but will not modify the previously parsed, or default, date and time.
+--
+-- A v'Format_Month', v'Format_DayYear', v'Format_Day' and v'Format_Tz_Offset'
+-- will check that the parsed number is within bounds. However,
+-- 'localTimeParseE' does not check that any resulting date or time is a valid
+-- one.
 localTimeParseE ::
      TimeFormat format
   => format -- ^ The format to use for parsing.
@@ -383,6 +433,11 @@ localTimeParseE fmt = loop ini fmtElems
       Left err         -> Left (x, err)
       Right (nacc, s') -> loop nacc xs s'
 
+  processOne ::
+       (DateTime, TimezoneOffset)
+    -> TimeFormatElem
+    -> [Char]
+    -> Either String ((DateTime, TimezoneOffset), [Char])
   processOne _   _               []     = Left "empty"
   processOne acc (Format_Text c) (x:xs)
     | c == x    = Right (acc, xs)
@@ -397,12 +452,42 @@ localTimeParseE fmt = loop ini fmtElems
             in  modDate (setYear year) acc
     )
     $ getNDigitNum 2 s
+  processOne acc Format_Month s =
+    let result = isNumber s :: Either String (Int, String)
+    in  case result of
+          Left err -> Left err
+          Right (m, s')
+            | m > 0 && m <= 12 ->
+                Right (modDate (setMonth (toEnum (m - 1))) acc, s')
+            | otherwise -> Left ("month invalid, got: " <> show m)
   processOne acc Format_Month2 s = onSuccess
     ( \m -> modDate (setMonth $ toEnum ((fromIntegral m - 1) `mod` 12)) acc
     )
     $ getNDigitNum 2 s
   processOne acc Format_MonthName_Short s =
     onSuccess (\m -> modDate (setMonth m) acc) $ getMonth s
+  processOne acc Format_DayYear s =
+    let y = (dateYear . dtDate . fst) acc
+        result = isNumber s :: Either String (Int, String)
+    in  case result of
+          Left err -> Left err
+          Right (d, s')
+              -- We can't be more helpful because we may not yet know the
+              -- intended year
+            | d > 0 && d <= 366 ->
+                let p = Period 0 0 (d - 1)
+                    startOfYear = Date y January 1
+                in  Right (modDate (const (dateAddPeriod startOfYear p)) acc, s')
+            | otherwise -> Left ("day of year invalid, got: " <> show d)
+  processOne acc Format_Day s =
+    let result = isNumber s :: Either String (Int, String)
+    in  case result of
+          Left err -> Left err
+          Right (d, s')
+              -- We can't be more helpful because we may not yet know the
+              -- intended month and year
+            | d > 0 && d <= 31 -> Right (modDate (setDay d) acc, s')
+            | otherwise -> Left ("day of month invalid, got: " <> show d)
   processOne acc Format_Day2 s =
     onSuccess (\d -> modDate (setDay d) acc) $ getNDigitNum 2 s
   processOne acc Format_Hour s =
@@ -419,6 +504,9 @@ localTimeParseE fmt = loop ini fmtElems
     onSuccess (\ns -> modTime (setNsMask (0,3) ns) acc) $ getNDigitNum 3 s
   processOne acc (Format_Precision p) s =
     onSuccess (\num -> modTime (setNS num) acc) $ getNDigitNum p s
+  processOne acc Format_TimezoneName s = case break isSpace s of
+    ("", _) -> Left ("no non-white space at start of: " <> s)
+    (_, s2) -> Right (acc, s2)
   processOne acc Format_UnixSecond s =
     onSuccess (\sec ->
       let newDate =
@@ -431,10 +519,27 @@ localTimeParseE fmt = loop ini fmtElems
     parseHMSign True acc c s
   processOne acc Format_TzHM (c:s) =
     parseHMSign False acc c s
-
-  processOne acc Format_Spaces (' ':s) = Right (acc, s)
-  -- catch all for unimplemented format.
-  processOne _ f _ = error ("unimplemented parsing format: " ++ show f)
+  processOne acc Format_Tz_Offset s@(c:cs) = case c of
+    '-' -> process True (12 * 60) cs
+    '+' -> process False (14 * 60) cs
+    _ -> process False (14 * 60) s
+   where
+    (dt, _) = acc
+    process isNeg limit s' =
+      let result = isNumber s' :: Either String (Int, String)
+      in  case result of
+            Left err -> Left err
+            Right (mins, rest)
+              | mins >= 0 && mins <= limit ->
+                  let mins' = if isNeg then negate mins else mins
+                  in Right ((dt, TimezoneOffset mins'), rest)
+              | otherwise ->
+                let sign = if isNeg then "-" else "+"
+                in Left ("offset invalid, got: " <> sign <> show mins)
+  processOne acc Format_Spaces s = case span isSpace s of
+    ("", _) -> Left ("no white space at start of: " <> s)
+    (_, s2) -> Right (acc, s2)
+  processOne acc (Format_Fct tff) s = timeFormatParse tff acc s
 
   parseHMSign expectColon acc signChar afterSign =
     case signChar of
@@ -516,13 +621,10 @@ localTimeParseE fmt = loop ini fmtElems
     s = 10 ^ shift
     m = 10 ^ mask
 
--- | Given the specified format, try to parse the given string as time value.
+-- | Like 'localTimeParseE', but with simpler handing of failure. Does not yield
+-- the remaining unparsed string on success.
 --
--- On failure, returns 'Nothing'.
---
--- If successful, yields 'Just' the parsed value.
---
--- For more elaborate needs use 'localTimeParseE'.
+-- On failure, returns 'Nothing'. If successful, yields 'Just' the parsed value.
 localTimeParse ::
      TimeFormat format
   => format -- ^ The format to use for parsing.